+++ RELEASES.md
@@ -1549,3 +1549,13 @@ Based on:
 - [typescript v0.8.0-alpha.1] .
 ### Releases
 - [NPM v0.8.0-alpha.1] https://www.npmjs.com/package/cribl-control-plane/v/0.8.0-alpha.1 - .
+
+## 2026-05-06 02:53:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.8.0-alpha.2] .
+### Releases
+- [NPM v0.8.0-alpha.2] https://www.npmjs.com/package/cribl-control-plane/v/0.8.0-alpha.2 - .